refactor: Factor out rendering checks
