+++ source/common/modules/markdown-editor/plugins/util/click-and-clear.ts
